+++ changelog.md
@@ -4,6 +4,7 @@
 * 'fromWords' and 'cloneFromWords' no longer take a number of bits.
 * Remove 'pad' and 'padWith'.
 * Remove 'any', 'anyBits', 'all', 'allBits', 'findIndex'.
+* Remove 'mapMInPlaceWithIndex', 'mapInPlaceWithIndex', 'mapMInPlace', 'mapInPlace'.
 
 # 0.2.0.1